requirements.txt
playwright==1.*
openpyxl==3.*
python-dotenv==1.*
keyring==24.*
